refactor(interface): add class interface and dataclass
